pyproject.toml
[project]
name = "sciencetldr"
version = "0.1.0"
description = "Automation pipeline for the Science TLDR podcast"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39",
    "lxml>=5.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "mutagen>=1.47",
    "python-slugify>=8.0",
    "markdown>=3.6",
]

[tool.uv]
package = false
